Correctly disable tests on Linux
The SP800-38G tests fail to execute on Linux, due to the inability to
set environment variables documented in a dotnet runtime issue.

This uses RuntimeInformation to conditionally exclude the test from
being run on Linux.

Switch from SupportedOSPlatform to RuntimeInformation.IsOSPlatform for TestRsaInterop

Guard this by a NET 6.0 requirement so that it doesn't affect old
Windows-only .NET framework versions.

Note: this only appears to fail on Linux when building from the CLI:

> dotnet test -v:normal -l \"console;verbosity=detailed\"

Building from an IDE (e.g., Rider) does not seem to be affected.

Signed-off-by: Alexander Scheel <alexander.scheel@keyfactor.com>