Implement the LIBSECURE Xtea cryptography library.
Implement the LIBSECURE XTEA cryptography library used on PlayStation devices.

This uses custom XTEA designed to be faster than the classic version.

The Number of rounds by default is 16, but in theory this can be adjusted up to 64.

NOTA: the number of rounds must be an even number! (actual library issue).