Embed picnic data in source to make it trim-friendly
Move the constant data for picnic from the bz2 files embedded as resources in the
assembly to be embedded as static properties in the source. Those properties can
then be analysed and trimmed away by the trimmer when unused.

In a test app which generates an ECDH key pair (i.e. does not use picnic), this
results in the following changes to the size of BouncyCastle.Cryptography.dll:

|           |  before |   after |
| --------- | ------- | ------- |
| untrimmed | 6989 KB | 6975 KB |
|   trimmed | 3993 KB | 2791 KB |

That is, the untrimmed assembly size is unaffected with this change, but the
trimmed assembly is reduced by 1202 KB which is exactly the size of the bz2 files.

Of the remaining 2791 KB, 2229 KB is the sike bz2 data, which is not treated here.

The source code was generated by reflecting over instances of the existing classes:

  using Org.BouncyCastle.Pqc.Crypto.Picnic;
  using System;
  using System.CodeDom.Compiler;
  using System.IO;
  using System.Linq;
  using System.Reflection;

  LowmcConstantsL5 lc = new();

  FieldInfo[] fields = lc.GetType().GetFields(BindingFlags.NonPublic | BindingFlags.Instance);

  using StreamWriter sw = new(@"C:\tmp\LowmcConstants.txt");
  using IndentedTextWriter tw = new(sw);

  tw.Indent++; //namespace
  tw.Indent++; //class
  tw.WriteLine(); // to get tabs on next line

  foreach (FieldInfo uintField in fields.Where(f => f.FieldType == typeof(uint[])))
  {
      uint[] data = (uint[])uintField.GetValue(lc);

      int nonZeroDataLength = data.AsSpan().LastIndexOfAnyExcept(0u) + 1;

      tw.WriteLine($"/// <summary>Length: {nonZeroDataLength}</summary>");
      tw.WriteLine($"private static readonly uint[] s_{uintField.Name} = new uint[]");
      tw.WriteLine("{");
      tw.Indent++;

      foreach (uint[] chunk in data.Take(nonZeroDataLength).Chunk(8))
      {
          tw.Write("0x");
          tw.Write(string.Join(", 0x", chunk.Select(u => u.ToString("X8"))));
          tw.WriteLine(",");
      }

      tw.Indent--;
      tw.WriteLine("};");
      tw.WriteLine();
  }