Added netstandard2.1 to the build targets
There are a number of APIs in BouncyCastle that are conditionally enabled
(such as APIs that make use of `ReadOnlySpan<T>`) when
`#if NETSTANDARD2_1_OR_GREATER` evaluates to `true`.

This causes issues if any library that references BouncyCastle targets netstandard2.1
(and would therefore build against the netstandard2.0 version of BC) is, itself,
referenced by an app (or other library) that targets net6.0+, resulting an a
"missing implementation" error.

Fixes issue #447 (at least in a practical sense)